Renames fixtures to mock tile data
